feat: wrap `Main` and `Auth `in navigations `AppNavigation`
