+++ src/navigation/AuthNavigation/AuthNavigation.tsx
@@ -1,0 +1,12 @@
+import { View } from "react-native";
+import { Text } from "react-native-paper";
+
+const AuthNavigation = () => {
+  return (
+    <View>
+      <Text>AuthNavigation</Text>
+    </View>
+  );
+};
+
+export default AuthNavigation;
